#105 chore: 픽스쳐 클래스 네이밍 인수 테스트로 변경
