+++ backend/src/test/java/moheng/acceptance/RecommendTripAcceptenceTest.java
@@ -1,16 +1,13 @@
 package moheng.acceptance;
 
-import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import moheng.acceptance.config.AcceptanceTestConfig;
 import moheng.auth.dto.AccessTokenResponse;
-import moheng.recommendtrip.dto.RecommendTripCreateRequest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.http.MediaType;
 
-import static moheng.acceptance.fixture.RecommendTripFixture.선호_여행지를_선택한다;
+import static moheng.acceptance.fixture.RecommendTripAcceptenceFixture.선호_여행지를_선택한다;
 import static moheng.acceptance.fixture.AuthAcceptanceFixture.자체_토큰을_생성한다;
 import static moheng.acceptance.fixture.TripAcceptenceFixture.여행지를_생성한다;
