+++ backend/src/test/java/moheng/acceptance/TripAcceptenceTest.java
@@ -1,10 +1,10 @@
 package moheng.acceptance;
 
-import static moheng.acceptance.fixture.KeywordFixture.키워드를_생성한다;
-import static moheng.acceptance.fixture.TripKeywordFixture.*;
+import static moheng.acceptance.fixture.KeywordAcceptenceFixture.키워드를_생성한다;
+import static moheng.acceptance.fixture.TripKeywordAcceptenceFixture.*;
 import static moheng.acceptance.fixture.TripAcceptenceFixture.*;
-import static moheng.acceptance.fixture.HttpStatus.상태코드_200이_반환된다;
-import static moheng.acceptance.fixture.HttpStatus.상태코드_204이_반환된다;
+import static moheng.acceptance.fixture.HttpStatusAcceptenceFixture.상태코드_200이_반환된다;
+import static moheng.acceptance.fixture.HttpStatusAcceptenceFixture.상태코드_204이_반환된다;
 import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.assertj.core.api.Assertions.assertThat;
