+++ backend/src/test/java/moheng/acceptance/fixture/RecommendTripAcceptenceFixture.java
@@ -3,16 +3,16 @@ package moheng.acceptance.fixture;
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-import moheng.trip.dto.TripCreateRequest;
-import moheng.trip.dto.TripKeywordCreateRequest;
+import moheng.recommendtrip.dto.RecommendTripCreateRequest;
 import org.springframework.http.MediaType;
 
-public class TripKeywordFixture {
-    public static ExtractableResponse<Response> 여행지_키워드를_생성한다(long keywordId, long tripId) {
+public class RecommendTripAcceptenceFixture {
+    public static ExtractableResponse<Response> 선호_여행지를_선택한다(final long tripId, final String accessToken) {
         return RestAssured.given().log().all()
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .body(new TripKeywordCreateRequest(keywordId, tripId))
-                .when().post("/keyword/trip")
+                .auth().oauth2(accessToken)
+                .body(new RecommendTripCreateRequest(tripId))
+                .when().post("/recommend")
                 .then().log().all()
                 .statusCode(org.springframework.http.HttpStatus.NO_CONTENT.value())
                 .extract();
