+++ backend/src/test/java/moheng/acceptance/fixture/TripKeywordAcceptenceFixture.java
@@ -3,17 +3,15 @@ package moheng.acceptance.fixture;
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-import moheng.recommendtrip.dto.RecommendTripCreateRequest;
-import moheng.trip.dto.TripCreateRequest;
+import moheng.trip.dto.TripKeywordCreateRequest;
 import org.springframework.http.MediaType;
 
-public class RecommendTripFixture {
-    public static ExtractableResponse<Response> 선호_여행지를_선택한다(final long tripId, final String accessToken) {
+public class TripKeywordAcceptenceFixture {
+    public static ExtractableResponse<Response> 여행지_키워드를_생성한다(long keywordId, long tripId) {
         return RestAssured.given().log().all()
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .auth().oauth2(accessToken)
-                .body(new RecommendTripCreateRequest(tripId))
-                .when().post("/recommend")
+                .body(new TripKeywordCreateRequest(keywordId, tripId))
+                .when().post("/keyword/trip")
                 .then().log().all()
                 .statusCode(org.springframework.http.HttpStatus.NO_CONTENT.value())
                 .extract();
